chore(dependencytrack): clean up code
* break down to a client configuration
* fix procken values for endpoint in helm chart for fqdns
